backend/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
pandas>=2.2
openpyxl>=3.1
flasgger>=0.9
flask-cors
psycopg2-binary>=2.9
python-dotenv>=1.0
alembic>=1.10
sqlalchemy>=2.0
PyJWT>=2.8
gunicorn>=21.2.0
requests>=2.31
jmespath>=1.0
python-dateutil>=2.8
pytest>=8.0
pytest-flask>=1.3

